snippets/requirements.txt
azure-functions
azure-search-documents>=11.5.0
openai>=1.40.0
azure-identity
requests
